// IBE-Frontend/src/interfaces/IExchangeRate.ts
import { translationInterface } from "./ITranslation";

export interface IExchangeRate{
    [key:string]:number;
}

export interface ICombinedStateGuests {
    [key: string]: number;
  }

  export interface ITenantConfiguration {
    tenantId:         string;
    tenantName:       string;
    tenantLogo:       string;
    backgroundImage:  string; 
    properties:       string[];
}

// Generated by https://quicktype.io

export interface IPropertyConfiguration {
    fields: Field[];
    guests: Guest[];
    rooms: number;
    maxPeoplePerRoom: number;
    maxLengthOfStay: number;
    filterTypes: FilterType[];
    bedType: BedType[];
    roomType: RoomType[];
    price: Price[];
    roomTypes: { [key: string]: RoomTypeInfo };
}

export interface Field {
    title: string;
    active: boolean;
}

export interface Guest {
    name: string;
    age: string;
    active: boolean;
}

export interface FilterType {
    title: string;
    active: boolean;
}

export interface BedType {
    name: string;
    active: boolean;
}

export interface RoomType {
    name: string;
    active: boolean;
}

export interface Price {
    name: string;
    active: boolean;
}

export interface RoomTypeInfo {
    images: string[];
}


// Generated by https://quicktype.io

export interface IPropertiesList {
    data: Data;
}

export interface Data {
    listProperties: ListProperty[];
}

export interface ListProperty {
    property_id:   number;
    property_name: string;
}

export interface ILandingState{
    language:string;
    currency:string;
    rate:IExchangeRate;
    startDateString:string | null;
    endDateString:string | null;
    state: "fulfilled" | "pending" | "error";
    startDate:Date | null,
    endDate:Date | null,
    isCalendarOpen:boolean;
    nightlyRates:any;
    totalGuestCount:number;
    initialCombinedStateGuests:ICombinedStateGuests;
    initialRoomCount:number;
    initialProperties:string[];
    initialTenantConfiguration:ITenantConfiguration;
    translations:translationInterface;
    tenantConfigurationsLoadingState:boolean;
    propertiesList:IPropertiesList;
    initialPropertyConfiguration:IPropertyConfiguration;
    accesibleCheckbox:boolean;
    propertyDropDown:boolean;
    showCalendarDropDown:boolean;
    guestDropDown:boolean;
    roomsDropDown:boolean;
    properyID:string;
}