// IBE-Frontend/src/interfaces/ITranslation.ts
export interface translationInterface {
  [key: string]: {
    subheading: string;
    myBookings: string;
    en: string;
    it: string;
    usd: string;
    inr: string;
    eur: string;
    login: string;
    price: string;
    propertyName: string;
    allProperties: string;
    selectRates: string;
    selectDates: string;
    checkIn: string;
    checkOut: string;
    guests: string;
    rooms: string;
    access: string;
    search: string;
    footer: string;
    applyDates: string;
    datemsg: string;
    day: string;
    from: string;
    night: string;
    roomResult: string;
    minimumAdult: string;
    maximumPeople: string;
    narrow: string;
    chooseRoom: string;
    chooseAddon: string;
    checkout: string;
    beds: string;
    checkInBetween: string;
    checkOutBetween: string;
    searchDates: string;
    king: string;
    queen: string;
    single: string;
    double: string;
    inclusive: string;
    specialDeal: string;
    seniorCitizenDiscount: string;
    selectRoom: string;
    showing: string;
    of: string;
    result: string;
    type: string;
    nearCityCenter: string;
    high: string;
    low: string;
    none: string;
    selectedPackege: string;
    errorInDates: string;
    noRoomsFound: string;
    newProperty: string;
    reviews: string;
    enterPromoCode: string;
    apply: string;
    standardRate: string;
    dealsAndPackages: string;
    amenities: string;
    tripItinerary: string;
    remove: string;
    CHECKOUT: string;
    continueShopping: string;
    feedback: string;
    travelerInfo: string;
    billingInfo: string;
    paymentInfo: string;
    next: string;
    needHelp: string;
    callNumber: string;
    officeHours: string;
    subtotal: string;
    taxesFees: string;
    vat: string;
    dueNow: string;
    dueAtResort: string;
    yourFeedback: string;
    submit: string;
    writeReview: string;
    thankYouFeedback: string;
    showIt: string;
    notApplicable: string;
    firstName: string;
    lastName: string;
    phone: string;
    email: string;
    nextBillingInfo: string;
    requiredField: string;
    numberTenDigits: string;
    zipCode: string;
    invalid: string;
    country: string;
    state: string;
    city: string;
    editTravellerInfo: string;
    nextPaymentInfo: string;
    cardNumber: string;
    expirationMonth: string;
    expirationYear: string;
    cardExpired: string;
    agreeTermsPolicies: string;
    sendSpecialOffers: string;
    iAgreeTo: string;
    termsPolicies: string;
    ofTravel: string;
    totalDue: string;
    editBillingInfo: string;
    purchase: string;
    total: string;
    specialPromoname: string;
    roomBreakdown: string;
    name:string;
    nightlyRate: string;
    taxesFeesPerNight: string;
    resortFee: string;
    occupanceTax: string;
    packageTotal: string;
    enterOtpCancelRoomBooking: string;
    enterOtp: string;
    confirmOtp: string;
    termsConditions: string;
    checkInCheckOut: string;
    checkOutDateChange: string;
    firstNightDeposit: string;
    nonrefundableCancellation: string;
    selfParking: string;
    estimatedTaxes: string;
    creditCardDeposit: string;
    upcomingReservation: string;
    spendTenEarnHundredFifty: string;
    print: string;
    cancellationPolicy: string;
    roomTotalSummary: string;
    totalForStay: string;
    address: string;
    noBookings: string;
    bookingDetails: string;
    bookingId: string;
    roomType: string;
    status: string;
    active: string;
    cancelled: string;
    totalCost: string;
    guestInfo: string;
    bookingCancelled: string;
    cancelRoom: string;
    timeLeft: string;
    roomsLeft: string;
    addToCompare: string;
    roomComparisonResult: string;
    closeCompare: string;
    clearComparison: string;
    averageRatings: string;
    anonymousRatings: string;
    wantAssistClickToChat: string;
    clickToHide: string;
    compare: string;
    hurry: string; 
    doubleBed: string;
    singleBed: string;
    area: string;
  };
}